🔧 Misc changes

+++ src/main/java/com/tiviacz/travelersbackpack/inventory/BackpackContainer.java
@@ -19,14 +19,14 @@ import org.jetbrains.annotations.Nullable;
 public class BackpackContainer implements MenuProvider, Nameable {
     public final ItemStack stack;
     public final Player player;
-    public final byte screenID;
+    public final int screenID;
     public final int index;
 
-    public BackpackContainer(ItemStack stack, Player player, byte screenID) {
+    public BackpackContainer(ItemStack stack, Player player, int screenID) {
         this(stack, player, screenID, -1);
     }
 
-    public BackpackContainer(ItemStack stack, Player player, byte screenID, int index) {
+    public BackpackContainer(ItemStack stack, Player player, int screenID, int index) {
         this.stack = stack;
         this.player = player;
         this.screenID = screenID;
@@ -53,33 +53,33 @@ public class BackpackContainer implements MenuProvider, Nameable {
         }
     }
 
-    public static FriendlyByteBuf saveExtraData(FriendlyByteBuf buf, @Nullable Player target, byte screenID) {
-        buf.writeByte(screenID);
+    public static FriendlyByteBuf saveExtraData(FriendlyByteBuf buf, @Nullable Player target, int screenID) {
+        buf.writeInt(screenID);
         buf.writeInt(target == null ? -1 : target.getId());
         return buf;
     }
 
-    public static FriendlyByteBuf saveExtraData(FriendlyByteBuf buf, int index, byte screenID) {
-        buf.writeByte(screenID);
+    public static FriendlyByteBuf saveExtraData(FriendlyByteBuf buf, int index, int screenID) {
+        buf.writeInt(screenID);
         buf.writeInt(index);
         return buf;
     }
 
     //Capability
-    public static void openBackpack(ServerPlayer serverPlayerEntity, ItemStack stack, byte screenID) {
+    public static void openBackpack(ServerPlayer serverPlayerEntity, ItemStack stack, int screenID) {
         if(!serverPlayerEntity.level().isClientSide) {
             serverPlayerEntity.openMenu(new BackpackContainer(stack, serverPlayerEntity, screenID), buf -> saveExtraData(buf, null, screenID));
         }
     }
 
     //Item
-    public static void openBackpack(ServerPlayer serverPlayerEntity, ItemStack stack, byte screenID, int index) {
+    public static void openBackpack(ServerPlayer serverPlayerEntity, ItemStack stack, int screenID, int index) {
         if(!serverPlayerEntity.level().isClientSide) {
             serverPlayerEntity.openMenu(new BackpackContainer(stack, serverPlayerEntity, screenID, index), buf -> saveExtraData(buf, index, screenID));
         }
     }
 
-    public static void openAnotherPlayerBackpack(ServerPlayer opener, ServerPlayer targetPlayer, ItemStack stack, byte screenID) {
+    public static void openAnotherPlayerBackpack(ServerPlayer opener, ServerPlayer targetPlayer, ItemStack stack, int screenID) {
         if(!opener.level().isClientSide) {
             synchroniseToOpener(opener, targetPlayer);
             opener.openMenu(new BackpackContainer(stack, targetPlayer, screenID), buf -> saveExtraData(buf, targetPlayer, screenID));
