+++ src/main/java/com/tiviacz/travelersbackpack/inventory/BackpackSettingsContainer.java
@@ -18,10 +18,10 @@ import org.jetbrains.annotations.Nullable;
 public class BackpackSettingsContainer implements MenuProvider, Nameable {
     public final ItemStack stack;
     public final Player player;
-    public final byte screenID;
+    public final int screenID;
     public final int index;
 
-    public BackpackSettingsContainer(ItemStack stack, Player player, byte screenID, int index) {
+    public BackpackSettingsContainer(ItemStack stack, Player player, int screenID, int index) {
         this.stack = stack;
         this.player = player;
         this.screenID = screenID;
@@ -38,9 +38,9 @@ public class BackpackSettingsContainer implements MenuProvider, Nameable {
         return Component.translatable("screen.travelersbackpack.item");
     }
 
-    public static RegistryFriendlyByteBuf saveSettingsExtraData(RegistryFriendlyByteBuf buf, byte screenID, int index) {
+    public static RegistryFriendlyByteBuf saveSettingsExtraData(RegistryFriendlyByteBuf buf, int screenID, int index) {
         buf.writeBoolean(false);
-        buf.writeByte(screenID);
+        buf.writeInt(screenID);
         buf.writeBlockPos(BlockPos.ZERO); //Not used
         buf.writeInt(index);
         return buf;
@@ -56,7 +56,7 @@ public class BackpackSettingsContainer implements MenuProvider, Nameable {
         }
     }
 
-    public static void openSettings(ServerPlayer serverPlayerEntity, ItemStack stack, byte screenID, int index) {
+    public static void openSettings(ServerPlayer serverPlayerEntity, ItemStack stack, int screenID, int index) {
         if(!serverPlayerEntity.level().isClientSide) {
             serverPlayerEntity.openMenu(new BackpackSettingsContainer(stack, serverPlayerEntity, screenID, index), buf -> saveSettingsExtraData(buf, screenID, index));
         }
